Add protocol files to files that generate false positives
This ensures that the `refresh_all` script does not add any data
from unrelated protocols.

I'd rather match on the "model" of the device and skip anything
that's not that model, but let's go for this first.

I'm not sure if this is the preferred way to prevent this from
happening, but at least it works :).

I skipped over any protocol that gave more or less lines of the
same signal -- not sure what to do about those.